# src/main/resources/swagger/api-animal.yaml
swagger: '2.0'

info:
  version: '0.0.1-SNAPSHOT'
  title: 'API - Animals'
  description: 'API de recursos de manipulação para Carros'

host: localhost:8080
basePath: /api/v1

schemes:
  - http
  - https

produces:
  - "application/json"
tags:
  - name: "Carros"
    description: "Operações para o recurso Carro"

paths:
  /carros:
    get:
      tags:
        - "Carros"
      summary: Lista todos os carros cadastrados
      description: Lista todos os detalhes dos carros cadastrados.
      operationId: listaTodosCarros
      produces:
        - application/json
      responses:
        200:
          description: "Representação dos dados de todos os carros cadastrados."
          schema:
            $ref: '#/definitions/ResponseGetCarros'
        404:
          description: "Nenhum carro encontrado"
        500:
          description: "Erro interno"
    post:
      tags:
      - "Carros"
      summary: "Adiciona um novo carro"
      description: "Inclui um novo carro retornando a URI de acesso ao recurso"
      operationId: "adicionaCarro"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto do tipo carro que será adicionado"
        required: true
        schema:
          $ref: "#/definitions/CarroModel"
      responses:
        201:
          description: "Carro adicionado com sucesso"
          schema:
            $ref: '#/definitions/ResponsePostCarros'
        405:
          description: "Entrada inválida"
        500:
          description: "Erro interno"
  /carros/{id}:
    get:
      tags:
        - "Carros"
      summary: Lista um carro específico
      description: Lista os detalhes de um carro específico de acordo com o ID.
      operationId: listCarroPorId
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: integer
          format: "int64"
          required: true
      responses:
        200:
          description: "Detalhe do carro encontrado"
          schema:
            $ref: '#/definitions/ResponseGetCarrosById'
        404:
          description: Nenhum carro encontrado correspondente ao ID informado
        500:
          description: Erro interno
    put:
      tags:
      - "Carros"
      summary: "Atualiza um carro existente."
      description: "Atualiza um novo carro retornando a URI de acesso ao recurso"
      operationId: "atualizaCarro"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: id
        in: path
        type: integer
        format: "int64"
        required: true
      - in: "body"
        name: "body"
        description: "Objeto do tipo carro que será atualizado"
        required: true
        schema:
          $ref: "#/definitions/CarroModel"
      responses:
        200:
          description: "Carro atualizado com sucesso"
          schema:
            $ref: '#/definitions/ResponsePutCarros'
        405:
          description: "Entrada inválida"
        500:
          description: "Erro interno"
    delete:
      tags:
      - "Carros"
      summary: "Remove um carro existente"
      description: "Exclui um novo carro retornando a URI de acesso ao recurso. Nesse caso, ao utiliar a URI será retornado uma resposta 404"
      operationId: "removeCarro"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - name: id
          in: path
          type: integer
          format: "int64"
          required: true
      responses:
        200:
          description: "Carro removido com sucesso"
          schema:
            $ref: '#/definitions/ResponseDeleteCarros'
        405:
          description: "Entrada inválida"
        500:
          description: "Erro interno"
definitions:
  ResponseGetCarros:
    type: object
    properties:
      carros:
        type: array
        items:
          $ref: '#/definitions/CarroModel'

  ResponseGetCarrosById:
    type: object
    properties:
      carro:
        $ref: '#/definitions/CarroModel'

  ResponsePostCarros:
    type: object
    properties:
      uri:
        type: "string"
      carro:
        $ref: '#/definitions/CarroModel'

  ResponsePutCarros:
    type: object
    properties:
      uri:
        type: "string"
      carro:
        $ref: '#/definitions/CarroModel'

  ResponseDeleteCarros:
    type: object
    properties:
      uri:
        type: "string"
      carro:
        $ref: '#/definitions/CarroModel'

  TipoCarro:
    type: "object"
    required:
      - "codigo"
      - "descricao"
    properties:
      id:
        type: "integer"
        format: "int64"
      codigo:
        type: "integer"
        format: "int64"
      descricao:
        type: "string"

  Carro:
    type: "object"
    required:
      - "id"
      - "codigo"
      - "descricao"
      - "tipo"
    properties:
      id:
        type: "integer"
        format: "int64"
      codigo:
        type: "integer"
        format: "int64"
      descricao:
        type: "string"
      dataCriacao:
        type: "string"
        format: "date-time"
      dataAtualizacao:
        type: "string"
        format: "date-time"
      tipo:
        $ref: "#/definitions/TipoCarro"
  CarroModel:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      codigo:
        type: "integer"
        format: "int64"
      descricao:
        type: "string"
      dataCriacao:
        type: "string"
      dataAtualizacao:
        type: "string"
      tipo:
        type: "integer"
        format: "int64"